HBA1C decimal

+++ cariro_implant_academy/lib/features/patientsMedical/medicalExamination/domain/entities/hba1cEntity.dart
@@ -5,7 +5,7 @@ import '../../../../../Helpers/CIA_DateConverters.dart';
 
 class HbA1cEntity extends Equatable{
   DateTime? date;
-  int? reading;
+  double? reading;
 
   HbA1cEntity({this.date, this.reading});
